Added styles for each FrenListItem component. Still needs works
